feat: move the render config to config file
